Organize project structure and remove unused files
