+++ service/templates/_jobs_table.html
