+++ start.bat
@@ -18,4 +18,4 @@ for /f "tokens=2 delims=:" %%A in ('ipconfig ^| findstr /R /C:"IPv4 Address"') d
 set IP=%IP:~1%
 echo Server will be available at http://%IP%:%PORT%
 
-python flask_app.py
+ python service\flask_app.py
