--- worker.log
@@ -1,66 +1,0 @@
-/root/venv/lib/python3.12/site-packages/celery/platforms.py:841: SecurityWarning: You're running the worker with superuser privileges: this is
-absolutely not recommended!
-
-Please specify a different user using the --uid option.
-
-User information: uid=0 euid=0 gid=0 egid=0
-
-  warnings.warn(SecurityWarning(ROOT_DISCOURAGED.format(
- 
- -------------- celery@AAPvhWvmVEUgTEC v5.5.3 (immunity)
---- ***** ----- 
--- ******* ---- Linux-6.6.87.1-microsoft-standard-WSL2-x86_64-with-glibc2.39 2025-06-14 05:11:53
-- *** --- * --- 
-- ** ---------- [config]
-- ** ---------- .> app:         celery_app:0x748dd671ef00
-- ** ---------- .> transport:   redis://localhost:6379/0
-- ** ---------- .> results:     redis://localhost:6379/0
-- *** --- * --- .> concurrency: 20 (prefork)
--- ******* ---- .> task events: OFF (enable -E to monitor tasks in this worker)
---- ***** ----- 
- -------------- [queues]
-                .> celery           exchange=celery(direct) key=celery
-                
-
-[tasks]
-  . tasks.run_task
-
-[2025-06-14 05:11:53,986: CRITICAL/MainProcess] Unrecoverable error: PermissionError(13, 'Permission denied')
-Traceback (most recent call last):
-  File "/root/venv/lib/python3.12/site-packages/celery/worker/worker.py", line 203, in start
-    self.blueprint.start(self)
-  File "/root/venv/lib/python3.12/site-packages/celery/bootsteps.py", line 116, in start
-    step.start(parent)
-  File "/root/venv/lib/python3.12/site-packages/celery/bootsteps.py", line 365, in start
-    return self.obj.start()
-           ^^^^^^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/celery/concurrency/base.py", line 130, in start
-    self.on_start()
-  File "/root/venv/lib/python3.12/site-packages/celery/concurrency/prefork.py", line 109, in on_start
-    P = self._pool = Pool(processes=self.limit,
-                     ^^^^^^^^^^^^^^^^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/celery/concurrency/asynpool.py", line 473, in __init__
-    super().__init__(processes, *args, **kwargs)
-  File "/root/venv/lib/python3.12/site-packages/billiard/pool.py", line 1046, in __init__
-    self._create_worker_process(i)
-  File "/root/venv/lib/python3.12/site-packages/celery/concurrency/asynpool.py", line 491, in _create_worker_process
-    return super()._create_worker_process(i)
-           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/billiard/pool.py", line 1142, in _create_worker_process
-    on_ready_counter = self._ctx.Value('i')
-                       ^^^^^^^^^^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/billiard/context.py", line 179, in Value
-    return Value(typecode_or_type, *args, lock=lock,
-           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/billiard/sharedctypes.py", line 81, in Value
-    lock = ctx.RLock()
-           ^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/billiard/context.py", line 108, in RLock
-    return RLock(ctx=self.get_context())
-           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-  File "/root/venv/lib/python3.12/site-packages/billiard/synchronize.py", line 206, in __init__
-    SemLock.__init__(self, RECURSIVE_MUTEX, 1, 1, ctx=ctx)
-  File "/root/venv/lib/python3.12/site-packages/billiard/synchronize.py", line 70, in __init__
-    sl = self._semlock = _billiard.SemLock(
-                         ^^^^^^^^^^^^^^^^^^
-PermissionError: [Errno 13] Permission denied
